refactor test cases for common tests
