+++ tests/dat/actions/swift3zip/build.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+./compile.sh Hello
